REF: test_searchorted for PeriodIndex (#33040)
